supposedly fixes network loss in aws
vrouter-agent entrypoint script performs manipulations on dhclient
processes. it needs to replace a running process with another.
now the entrypoint process runs in a ct namespace thus the replacement
doesn't occur, we end with 2 dhclient processes which leads to a
network loss.
let's start vrouter-agent in host pid namespace to workaround.

related-jira-bug: CEM-19790
Change-Id: Id2c051f566b16c8928980ccb3438c5bbe08d7fde